requirements.txt
anthropic==0.25.6
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
coverage==7.7.1
distro==1.9.0
dj-database-url==2.1.0
Django>=4.2.16
django-allauth==65.5.0
django-cors-headers==4.3.1
django-discover-runner==1.0
django-oauth-toolkit==3.0.1
django-safedelete==1.3.1
django-webtest==1.9.11
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
exceptiongroup==1.2.0
factory-boy==3.3.0
Faker==12.0.1
filelock==3.13.4
fsspec==2024.3.1
google==3.0.0
google-api-python-client==2.164.0
google-auth==2.38.0
groq==0.19.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
legacy-cgi==2.6.2
mixer==7.2.2
mock==5.1.0
model-mommy==2.0.0
openai==1.14.2
packaging==24.0
pluggy==1.4.0
psycopg==3.1.18
psycopg2-binary==2.9.10
pydantic==1.10.2
PyJWT==2.8.0
pytest==8.1.1
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-social-auth==0.3.6
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
sentry-sdk==2.22.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.4.4
tokenizers==0.21.0
tomli==2.0.1
tqdm==4.66.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
waitress==3.0.0
webob==1.8.9
WebTest==3.0.0
whitenoise==6.6.0